Factor out patched methods from patching built-ins
